LPD-31502 Move to existing file
